Refator emit function to a diffrent file
